some restructure, removed model support, level packs

+++ Levels/AmbLight.cs
@@ -2,16 +2,16 @@ using System.Drawing;
 using System.Text.Json.Serialization;
 
 namespace ThemModdingHerds.Levels;
-public class AmbLight(byte red,byte green,byte blue) : ILight, IColor
+public class AmbLight(ushort red,ushort green,ushort blue) : ILight, IColor
 {
     [JsonPropertyName("type")]
     public string Type => "Amb";
     [JsonPropertyName("red")]
-    public byte Red {get; set;} = red;
+    public ushort Red {get; set;} = red;
     [JsonPropertyName("green")]
-    public byte Green {get; set;} = green;
+    public ushort Green {get; set;} = green;
     [JsonPropertyName("blue")]
-    public byte Blue {get; set;} = blue;
+    public ushort Blue {get; set;} = blue;
     [JsonIgnore]
     public Color Color {get => Color.FromArgb(0,Red,Green,Blue);set{Red = value.R;Green = value.G;Blue = value.B;}}
     public AmbLight(Color color): this(color.R,color.G,color.B)
@@ -28,7 +28,7 @@ public class AmbLight(byte red,byte green,byte blue) : ILight, IColor
     }
     public static AmbLight Parse(string s)
     {
-        byte[] values = LevelParsers.ParseSameMultiple<byte>(s,"Light: Amb",3);
+        ushort[] values = LevelParsers.ParseSameMultiple<ushort>(s,"Light: Amb",3);
         return new(values[0],values[1],values[2]);
     }
 }
