+++ Levels/IColor.cs
@@ -3,8 +3,8 @@ using System.Drawing;
 namespace ThemModdingHerds.Levels;
 public interface IColor
 {
-    public byte Red {get; set;}
-    public byte Green {get; set;}
-    public byte Blue {get; set;}
+    public ushort Red {get; set;}
+    public ushort Green {get; set;}
+    public ushort Blue {get; set;}
     public Color Color {get;set;}
 }
